Refactor OOP lab9-12

+++ object-oriented-programming/labs/lab9-12/mem_repository.cpp
@@ -31,17 +31,19 @@ MemRepository::~MemRepository()
 void MemRepository::deleteArray()
 {
     this->deleteInternalArray(array);
-    delete array; array = NULL;
+    delete array;
+    array = NULL;
 }
 
-void MemRepository::deleteInternalArray(vector<Ingredient*> *toDelete)
+void MemRepository::deleteInternalArray(vector<Ingredient*>* toDelete)
 {
-    for(unsigned int i = 0; i < toDelete->size(); i++)
+    for (unsigned int i = 0; i < toDelete->size(); i++)
     {
-        Ingredient *to_delete = toDelete->at(i);
-        if(to_delete != NULL)
+        Ingredient* to_delete = toDelete->at(i);
+        if (to_delete != NULL)
         {
-            delete to_delete; to_delete = NULL;
+            delete to_delete;
+            to_delete = NULL;
         }
         else
         {
@@ -59,21 +61,22 @@ void MemRepository::clear()
 void MemRepository::deleteUndo()
 {
     // free undo
-    for(unsigned int i = 0; i < undoVector->size(); i++)
+    for (unsigned int i = 0; i < undoVector->size(); i++)
     {
         this->deleteInternalArray(this->undoVector->at(i));
 
         delete undoVector->at(i);
     }
 
-    delete undoVector; undoVector = NULL;
+    delete undoVector;
+    undoVector = NULL;
 }
 
 int MemRepository::getIndexByID(const unsigned int id) const
 {
-    for(unsigned int i = 0; i < array->size(); i++)
+    for (unsigned int i = 0; i < array->size(); i++)
     {
-        if(array->at(i)->getId() == id)
+        if (array->at(i)->getId() == id)
         {
             return i;
         }
@@ -82,19 +85,13 @@ int MemRepository::getIndexByID(const unsigned int id) const
     return -1;
 }
 
-Ingredient* MemRepository::getById(const unsigned int id) const
-{
-    return array->at(this->getIndexByID(id));
-}
+Ingredient* MemRepository::getById(const unsigned int id) const { return array->at(this->getIndexByID(id)); }
 
-Ingredient* MemRepository::getByIndex(const unsigned int index) const
-{
-    return array->at(index);
-}
+Ingredient* MemRepository::getByIndex(const unsigned int index) const { return array->at(index); }
 
 bool MemRepository::exists(unsigned int id)
 {
-    if(this->getIndexByID(id) == -1)
+    if (this->getIndexByID(id) == -1)
     {
         return false;
     }
@@ -102,22 +99,19 @@ bool MemRepository::exists(unsigned int id)
     return true;
 }
 
-void MemRepository::addIngredient(Ingredient *ingredient)
+void MemRepository::addIngredient(Ingredient* ingredient)
 {
     this->addToUndo();
     array->push_back(ingredient);
 }
 
-unsigned int MemRepository::getLength() const
-{
-    return array->size();
-}
+unsigned int MemRepository::getLength() const { return array->size(); }
 
 void MemRepository::removeIngredient(unsigned int id)
 {
     this->addToUndo();
     int index = this->getIndexByID(id);
-    if(index == -1)
+    if (index == -1)
     {
         printDebug("Id does not exist");
         throw "Id does not exist";
@@ -133,10 +127,10 @@ void MemRepository::filterBy(const RepositoryFilter& filter)
     this->addToUndo();
     auto new_vector = new vector<Ingredient*>;
 
-    for(unsigned int i = 0; i < array->size(); i++)
+    for (unsigned int i = 0; i < array->size(); i++)
     {
-        Ingredient *ingredient = array->at(i);
-        if(filter.isValid(ingredient))
+        Ingredient* ingredient = array->at(i);
+        if (filter.isValid(ingredient))
         {
             new_vector->push_back(ingredient);
         }
@@ -160,11 +154,11 @@ void MemRepository::addToUndo()
 {
     auto undo_vector = new vector<Ingredient*>;
 
-    for(unsigned int i = 0; i < array->size(); i++)
+    for (unsigned int i = 0; i < array->size(); i++)
     {
-        Ingredient *old_ingredient = array->at(i);
+        Ingredient* old_ingredient = array->at(i);
         undo_vector->push_back(new Ingredient(old_ingredient->getId(), old_ingredient->getQuantity(),
-                                              old_ingredient->getName(), old_ingredient->getProducer()));
+            old_ingredient->getName(), old_ingredient->getProducer()));
     }
 
     this->undoVector->push_back(undo_vector);
@@ -172,7 +166,7 @@ void MemRepository::addToUndo()
 
 bool MemRepository::undo()
 {
-    if(undoVector->empty())
+    if (undoVector->empty())
     {
         return false;
     }
@@ -184,17 +178,16 @@ bool MemRepository::undo()
     return true;
 }
 
-
 void MemRepository::toString() const
 {
-    cout << setw(4) << "ID" << setw(15) << "Quantity"
-         << setw(30) << "Name" << setw(30) << "Producer" << endl << endl;
+    cout << setw(4) << "ID" << setw(15) << "Quantity" << setw(30) << "Name" << setw(30) << "Producer" << endl
+         << endl;
 
-    for(int i = 0; i < array->size(); i++)
+    for (size_t i = 0; i < array->size(); i++)
     {
-        Ingredient *it = array->at(i);
-        cout << setw(4) << it->getId() << setw(15) << it->getQuantity()
-             << setw(30) << it->getName() << setw(30) << it->getProducer() << endl << endl;
+        Ingredient* it = array->at(i);
+        cout << setw(4) << it->getId() << setw(15) << it->getQuantity() << setw(30) << it->getName() << setw(30)
+             << it->getProducer() << endl
+             << endl;
     }
 }
-
